Move everything to `MenuBuilder.svelte`
